Restructured the files for patterns in the matcher, added UnitState trait for Phone and NonBound, added comments
